fix:booth 좋아요 업데이트 수정 boothLikedRepository BoothLiked찾는 메서드 ì추가 boothLikedService 주석 수정 ãDetermineService 주석 수정 Booth  countLike Long -> int로 수정

+++ NSU_FESTIVAL/src/main/java/com/example/nsu_festival/domain/booth/dto/BoothDetailDto.java
@@ -20,7 +20,7 @@ public class BoothDetailDto {
     private Long boothId;
     private String title;
     private String content;
-    private Long countLike;
+    private int countLike;
     private String area;
     private List<BoothCategory> boothCategories;
     private BoothImage boothImage;
