+++ NSU_FESTIVAL/src/main/java/com/example/nsu_festival/domain/booth/entity/Booth.java
@@ -26,7 +26,7 @@ public class Booth {
     @NonNull
     private String content;
 
-    private Long countLike;
+    private int countLike;
 
     private String area;
